fix: rejigger svelte as a class & ditch removal
Fixes #522
