--- packages/svelte/src/style.js
@@ -1,5 +1,0 @@
-"use strict";
-
-module.exports = () => ({
-    code : "/* replaced by modular-css */",
-});
