+++ packages/svelte/svelte.js
@@ -1,28 +1,5 @@
 "use strict";
 
-const Processor = require("@modular-css/processor");
+const Preprocessor = require("./src/preprocessor.js");
 
-const markup = require("./src/markup.js");
-const style = require("./src/style.js");
-
-module.exports = function(args) {
-    const config = Object.assign(
-        Object.create(null),
-        {
-            strict  : false,
-            verbose : false,
-        },
-        args
-    );
-
-    const processor = new Processor(config);
-
-    return {
-        processor,
-
-        preprocess : {
-            markup : markup(processor, config),
-            style,
-        },
-    };
-};
+module.exports = (args) => new Preprocessor(args);
